requirements.txt
apscheduler==3.10.4
ffmpeg-python==0.2.0
flask==3.0.2
google-api-python-client==2.108.0
google-auth==2.33.0
google-auth-oauthlib==1.2.0
requests==2.32.3
selenium==4.13.0
appium-python-client==3.1.1

